+++ tsup.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   entry: {
     cli: "src/cli/main.ts",
     main: "src/main/main.ts",
+    "plugin-renderer": "src/plugin-renderer/renderer.ts",
     preload: "src/main/preload.ts",
   },
   outDir: "dist/main",